requirements.txt
appnope==0.1.3
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.5.2
attrs==21.4.0
aws-sam-translator~=1.28.1
aws-xray-sdk==2.9.0
backcall==0.2.0
beautifulsoup4==4.9.3
bleach~=5.0.1
blis==0.7.7
blog==1.0.1
boto3==1.24.12
botocore==1.27.12
cachetools==5.2.0
cachier==1.5.4
catalogue==2.0.7
certifi==2022.5.18.1
cffi==1.15.0
cfn-lint==0.61.0
charset-normalizer==2.0.12
click==8.1.0
colorama==0.4.5
coverage==6.3.2
cryptography==37.0.2
cssselect==1.1.0
cssutils==2.4.2
cymem==2.0.6
debugpy==1.6.0
decorator==5.1.1
defusedxml==0.7.1
dependency-injector~=4.39.1
Django==4.0.5
django-extensions==3.1.5
django-filter==21.1
django-modelcluster==6.0
django-permissionedforms==0.1
django-svg-image-form-field==1.0.1
django-taggit==2.1.0
django-treebeard==4.5.1
djangorestframework==3.13.1
docker==5.0.3
docopt==0.6.2
draftjs-exporter==2.1.7
ecdsa~=0.17.0
entrypoints==0.4
equipment==0.0.16
et-xmlfile==1.1.0
executing==0.8.3
Faker==13.3.4
fastjsonschema==2.15.3
future==0.18.2
futures==2.1.6
graphql-core==3.2.1
html2text==2020.1.16
html5lib~=1.1
icecream==2.1.2
idna==3.3
ipykernel==6.15.0
ipython==8.4.0
ipython-genutils==0.2.0
ipywidgets==7.7.1
jedi==0.18.1
jinja-markdown==1.210911
Jinja2==3.1.1
jmespath==1.0.1
jschema-to-python==1.2.3
jsondiff==2.0.0
jsonpatch==1.32
jsonpickle==2.2.0
jsonpointer==2.3
jsonschema~=3.2.0
junit-xml~=1.9
jupyter==1.0.0
jupyter-client==7.3.4
jupyter-console==6.4.4
jupyter-core==4.10.0
jupyterlab-pygments==0.2.2
jupyterlab-widgets==1.1.1
l18n~=2021.3
langcodes==3.3.0
lxml==4.9.0
Markdown==3.3.7
MarkupSafe==2.1.1
matplotlib-inline==0.1.3
mistune==0.8.4
moto==3.1.3
murmurhash==1.0.7
nbclient==0.6.6
nbconvert==6.5.0
nbformat==5.4.0
nest-asyncio==1.5.5
networkx==2.8.4
notebook==6.4.12
numpy==1.22.4
openpyxl==3.0.10
packaging==21.3
pandocfilters==1.5.0
parent-import==0.0.1
parso==0.8.3
pathtools==0.1.2
pathy==0.6.1
pbr==5.9.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.1.1
portalocker==2.4.0
premailer==3.10.0
preshed==3.0.6
prometheus-client==0.14.1
prompt-toolkit==3.0.30
psutil==5.9.1
ptyprocess==0.7.0
pure-eval==0.2.2
py-dto==1.0.0
pyasn1==0.4.8
pycparser==2.21
pydantic==1.8.2
pyembed==1.3.3
pyembed-markdown==1.1.0
Pygments==2.12.0
pymdown-extensions==9.5
pyparsing==3.0.9
pyrsistent==0.18.1
pystreams~=0.6
python-dateutil==2.8.2
python-dotenv==0.20.0
python-jose==3.3.0
python-optional==1.1.0
pytz==2022.1
PyYAML==6.0
pyzmq==23.2.0
qtconsole==5.3.1
QtPy==2.1.0
repoze.lru==0.6
requests==2.27.1
responses==0.21.0
rsa==4.8
runtype==0.2.4
s3transfer==0.6.0
sarif-om==1.0.4
schedule==1.1.0
Send2Trash==1.8.0
six~=1.16.0
smart-open==5.2.1
soupsieve==2.3.2.post1
spacy==3.3.1
spacy-legacy==3.0.9
spacy-loggers==1.0.2
sqlparse==0.4.2
srsly==2.4.3
sshpubkeys==3.3.1
stack-data==0.3.0
tablib==3.2.1
telepath==0.2
terminado==0.15.0
thinc==8.0.17
tinycss2==1.1.1
tornado==6.2
tqdm==4.64.0
traitlets==5.3.0
typing_extensions==4.2.0
Unidecode==1.3.4
urllib3==1.26.9
wagtail==3.0
wagtail-modelchooser==2.14.1
wasabi==0.9.1
watchdog==2.1.9
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.3.2
Werkzeug==2.1.2
widgetsnbextension==3.6.1
Willow==1.4.1
wrapt==1.14.1
xlrd==2.0.1
XlsxWriter==1.4.5
xlwt==1.3.0
xmltodict==0.13.0
